+++ packages/playground/apps/components/custom-frame-panel.ts
