+++ packages/playground/apps/components/custom-outline-panel.ts
