+++ packages/playground/apps/env.d.ts
@@ -1,12 +1,7 @@
 import type { TestUtils } from '@blocksuite/blocks';
 import type { EditorHost } from '@blocksuite/lit';
 import type { AffineEditorContainer } from '@blocksuite/presets';
-import type {
-  BlockSchema,
-  DocProvider,
-  Page,
-  Workspace,
-} from '@blocksuite/store';
+import type { BlockSchema, Page, Workspace } from '@blocksuite/store';
 import type { Job } from '@blocksuite/store';
 import type { z } from 'zod';
 
@@ -27,9 +22,6 @@ declare global {
     std: typeof std;
     testUtils: TestUtils;
     host: EditorHost;
-
-    // TODO: remove this when provider support subdocument
-    subdocProviders: Map<string, DocProvider[]>;
 
     devtoolsFormatters: {
       header: (obj: unknown, config: unknown) => null | HTMLTemplate;
