+++ packages/playground/apps/providers/async-call-rpc.ts
@@ -49,7 +49,9 @@ export function createAsyncCallRPCProvider(
 
   const impl = {
     queryDocState: async (guid, targetClientId) => {
+      console.log('queryDocState', guid, targetClientId);
       const doc = docMap.get(guid);
+
       if (!doc) {
         return false;
       }
@@ -59,6 +61,7 @@ export function createAsyncCallRPCProvider(
       return Y.encodeStateAsUpdate(doc);
     },
     sendDocUpdate: async (guid, update) => {
+      console.log('sendDocUpdate', guid, update);
       const doc = docMap.get(guid);
       if (!doc) {
         // This case happens when the father doc is not yet updated,
@@ -234,7 +237,6 @@ export function createAsyncCallRPCProvider(
 
   let connected = false;
   const apis = {
-    passive: true,
     connect() {
       connected = true;
       registerDoc(doc);
